fix: resolve 344 clippy warnings (pedantic+nursery clean)
- 177 doc_markdown: backtick identifiers in doc comments across ~80 files
- ~98 cast warnings: targeted #[allow] on test functions, one real fix
  (f64::EPSILON as f32 → f32::EPSILON)
- 16 unfulfilled lint expectations removed
- 7 missing_errors_doc: added # Errors sections
- 5 unchecked_time_subtraction: switched to checked_sub().unwrap()
- 4 unused imports removed
- Misc: needless_return, bare URL, needless_collect suppression

Made-with: Cursor

+++ crates/beardog-utils/src/const_eval.rs
@@ -807,7 +807,7 @@ mod tests {
     fn test_const_tables_fast_sin_wraparound() {
         let sin_360 = ConstTables::fast_sin(360);
         let sin_0 = ConstTables::fast_sin(0);
-        assert!((sin_360 - sin_0).abs() < f64::EPSILON as f32);
+        assert!((sin_360 - sin_0).abs() < f32::EPSILON);
     }
 
     #[test]
